# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Manpower Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: Get all payments (admin) or my payments (crew)
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    put:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentRequest"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        crewId:
          type: integer
        crewName:
          type: string
        shiftClaimId:
          type: integer
        shiftRole:
          type: string
        eventTitle:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, processing, paid, failed]
        paymentMethod:
          type: string
        reference:
          type: string
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - crewId
        - crewName
        - amount
        - status
        - createdAt
    CreatePaymentRequest:
      type: object
      properties:
        crewId:
          type: integer
        shiftClaimId:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
        reference:
          type: string
        notes:
          type: string
      required:
        - crewId
        - amount
    UpdatePaymentRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, paid, failed]
        reference:
          type: string
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
...