# lib/api-zod/src/generated/types/updatePaymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Manpower Agency API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePaymentRequestStatus } from "./updatePaymentRequestStatus";

export interface UpdatePaymentRequest {
  status?: UpdatePaymentRequestStatus;
  reference?: string;
  notes?: string;
  paidAt?: Date;
}